FloydHub has shut down
First of all, thanks for this amazing resource! This is just what I needed.

I was looking through the list of companies, and it appears that FloydHub has shut down and is no longer accepting new posts.

https://blog.floydhub.com/floydhub-has-shut-down/ (see the P.S. at the end of the article)

I wasn't sure if you were keeping it on the list for any reason, but just wanted to let you know!

Thanks again!